Create Hand_Code_Using_Servos_and_Myoware_Muscle_Sensors.ino
this is the new code that will be the main code to be used with the muscle sensors. i changed some function name and added/ remove a struct from the template code.  change the template code to make sure we get a analog signal instead of a digital signal as inputs. code not yet tested